+++ tests/metadata/test_roundtrips.py
@@ -88,7 +88,6 @@ def test_optional_field_roundtrip(converter, cl_and_vals):
         a = typed(Optional[cl])
 
     inst = C(a=cl(*vals))
-
     assert inst == converter.structure(converter.unstructure(inst), C)
 
     inst = C(a=None)